+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Polarion Changelog
 
+## 0.0.7
+
+- Added GetDocumentSpacesAsync method
+- Refactored Client class into partial classes
+
 ## 0.0.6
 
 - Update PolarionClientConfiguration type to no longer need requred fields. This is to support more use cases in json serialization.